// apps/web/src/gen/api/admin/model/country.ts
/**
 * Admin
 * Admin service is responsible for User management and authentication. Also it provides static data i.e list of Skills and Countries 
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Country { 
    id: string;
    name: string;
    code: string;
}