// apps/web/src/gen/api/admin/model/createUserDto.ts
/**
 * Admin
 * Admin service is responsible for User management and authentication. Also it provides static data i.e list of Skills and Countries 
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateUserDto { 
    firstName: string;
    lastName: string;
    username: string;
    password: string;
    birthDate?: string;
    gender?: CreateUserDto.GenderEnum;
    experience?: number;
    roles: Array<CreateUserDto.RolesEnum>;
    country?: string;
    skills: Array<string>;
}
export namespace CreateUserDto {
    export type GenderEnum = 'male' | 'female';
    export const GenderEnum = {
        Male: 'male' as GenderEnum,
        Female: 'female' as GenderEnum
    };
    export type RolesEnum = 'admin' | 'developer' | 'crowdworker';
    export const RolesEnum = {
        Admin: 'admin' as RolesEnum,
        Developer: 'developer' as RolesEnum,
        Crowdworker: 'crowdworker' as RolesEnum
    };
}